Override global configuration on the namespace level (#10545)
* Block override configuration in namespace

* Block override configuration in namespace - tests

* Fix after review

+++ components/function-controller/internal/controllers/kubernetes/secret_service.go
@@ -4,6 +4,8 @@ import (
 	"context"
 	"fmt"
 
+	"github.com/kyma-project/kyma/components/function-controller/pkg/apis/serverless/v1alpha1"
+
 	"github.com/go-logr/logr"
 	corev1 "k8s.io/api/core/v1"
 	"k8s.io/apimachinery/pkg/api/errors"
@@ -59,7 +61,9 @@ func (r *secretService) UpdateNamespace(ctx context.Context, logger logr.Logger,
 		logger.Error(err, fmt.Sprintf("Gathering existing Secret '%s/%s' failed", namespace, baseInstance.GetName()))
 		return err
 	}
-
+	if instance.Labels[v1alpha1.FunctionManagedByLabel] == v1alpha1.FunctionResourceLabelUserValue {
+		return nil
+	}
 	return r.updateSecret(ctx, logger, instance, baseInstance)
 }
 
@@ -133,7 +137,9 @@ func (r *secretService) deleteSecret(ctx context.Context, logger logr.Logger, na
 	if err := r.client.Get(ctx, client.ObjectKey{Namespace: namespace, Name: baseInstanceName}, instance); err != nil {
 		return client.IgnoreNotFound(err)
 	}
-
+	if instance.Labels[v1alpha1.FunctionManagedByLabel] == v1alpha1.FunctionResourceLabelUserValue {
+		return nil
+	}
 	if err := r.client.Delete(ctx, instance); err != nil {
 		logger.Error(err, fmt.Sprintf("Deleting Secret '%s/%s' failed", namespace, baseInstanceName))
 		return err
